pyproject.toml
[project]
name = "observatory-context"
version = "0.1.0"
description = "OpenViking migration helpers for the BERIL research observatory"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.26.0",
    "openviking>=0.2.9",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.15.0",
]
berdl = [
    "boto3>=1.34.0",
    "matplotlib",
    "pandas",
    "pproxy",
]
ingest = [
    "jupyter",
    "linkml>=1.9.4",
    "linkml-runtime>=1.9.5",
    "linkml-validator>=0.4.5",
    "minio>=7.2.0",
    "nbconvert",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["observatory_context"]

[tool.pytest.ini_options]
testpaths = ["scripts/tests", "tests"]
